`crates/node` docs
