+++ crates/node/src/error.rs
@@ -1,25 +1,33 @@
 use thiserror::Error;
 
+/// Main error type
 #[derive(Error, Debug)]
 pub enum Error {
+    /// Database error ([`sqlx::Error`])
     #[error("Database error: {0}")]
     Database(#[from] sqlx::Error),
 
+    /// IO error
     #[error("IO error: {0}")]
     Io(#[from] std::io::Error),
 
+    /// Serialization error ([`serde_json::Error`])
     #[error("Serialization error: {0}")]
     Serialization(#[from] serde_json::Error),
 
+    /// gRPC error-status
     #[error("gRPC error: {0}")]
     GrpcStatus(Box<tonic::Status>),
 
+    /// gRPC connection error
     #[error("gRPC error: {0}")]
     GrpcTransport(#[from] tonic::transport::Error),
 
+    /// Internal node error
     #[error("Internal server error: {0}")]
     Internal(String),
 
+    /// Generic node error
     #[error("Error: {0}")]
     Generic(#[from] anyhow::Error),
 }
@@ -30,4 +38,5 @@ impl From<tonic::Status> for Error {
     }
 }
 
+/// Main result type
 pub type Result<T> = std::result::Result<T, Error>;
