+++ crates/node/src/types.rs
@@ -11,12 +11,17 @@ use serde::{Deserialize, Serialize};
 /// A note stored in the database
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct StoredNote {
+    /// Note header
     #[serde(
         serialize_with = "serialize_note_header",
         deserialize_with = "deserialize_note_header"
     )]
     pub header: NoteHeader,
+    /// Note details
+    ///
+    /// Can be encrypted.
     pub details: Vec<u8>,
+    /// Reference timestamp
     pub created_at: DateTime<Utc>,
 }
 
@@ -46,6 +51,7 @@ impl TryFrom<StoredNote> for miden_private_transport_proto::TransportNoteTimesta
     }
 }
 
+/// Helper converter from [`prost_types::Timestamp`] to `DateTime<Utc>`
 pub fn proto_timestamp_to_datetime(pts: prost_types::Timestamp) -> anyhow::Result<DateTime<Utc>> {
     let dts = DateTime::from_timestamp(
         pts.seconds,
@@ -78,45 +84,3 @@ where
     })
 }
 
-pub fn random_note_id() -> NoteId {
-    use miden_objects::{Felt, Word};
-    use rand::Rng;
-
-    let mut rng = rand::rng();
-
-    let recipient = Word::from([
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-    ]);
-    let asset_commitment = Word::from([
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-        Felt::new(rng.random::<u64>()),
-    ]);
-
-    NoteId::new(recipient, asset_commitment)
-}
-
-pub const TEST_TAG: u32 = 3_221_225_472;
-pub fn test_note_header() -> NoteHeader {
-    use miden_objects::{
-        Felt,
-        account::AccountId,
-        note::{NoteExecutionHint, NoteMetadata, NoteType},
-        testing::account_id::ACCOUNT_ID_MAX_ZEROES,
-    };
-
-    let id = random_note_id();
-    let sender = AccountId::try_from(ACCOUNT_ID_MAX_ZEROES).unwrap();
-    let note_type = NoteType::Private;
-    let tag = NoteTag::from_account_id(sender);
-    let aux = Felt::try_from(0xffff_ffff_0000_0000u64).unwrap();
-    let execution_hint = NoteExecutionHint::None;
-
-    let metadata = NoteMetadata::new(sender, note_type, tag, execution_hint, aux).unwrap();
-
-    NoteHeader::new(id, metadata)
-}
